feat: added inverse distance weighting (IDW) extrapolation
refactor: move ellipsoid and love number parameters to `earth` module
refactor: use `scipy.spatial.KDTree` as `cKDTree` is now an alias
