+++ doc/source/api_reference/constituents.rst
@@ -29,14 +29,6 @@ constituents
 
 .. autofunction:: pyTMD.constituents._constituent_parameters
 
-.. autofunction:: pyTMD.constituents._love_numbers
-
-.. autofunction:: pyTMD.constituents._complex_love_numbers
-
-.. autofunction:: pyTMD.constituents._degree_love_numbers
-
-.. autofunction:: pyTMD.constituents._melchior_table_52
-
 .. autofunction:: pyTMD.constituents._parse_tide_potential_table
 
 .. autofunction:: pyTMD.constituents._parse_rotation_rate_table
