+++ pyTMD/__init__.py
@@ -15,6 +15,7 @@ Documentation is available at https://pytmd.readthedocs.io
 import pyTMD.astro
 import pyTMD.compute
 import pyTMD.constituents
+import pyTMD.earth
 import pyTMD.ellipse
 import pyTMD.interpolate
 import pyTMD.math
