+++ pyTMD/constituents.py
@@ -26,6 +26,7 @@ REFERENCES:
 UPDATE HISTORY:
     Updated 05/2026: use numpy hypot function to calculate magnitudes
         deprecate minor table and arguments table functions
+        moved body tide Love/Shida numbers to earth module
     Updated 04/2026: add constituent mapping for TICON-4 sigma1 (SGM)
     Updated 03/2026: add degree-dependent body tide love number tables
     Updated 12/2025: added m1a and m1b constituents to nodal corrections
@@ -114,10 +115,6 @@ __all__ = [
     "aliasing_period",
     "_constituent_parameters",
     "_frequency",
-    "_love_numbers",
-    "_complex_love_numbers",
-    "_degree_love_numbers",
-    "_melchior_table_52",
     "_parse_tide_potential_table",
     "_parse_rotation_rate_table",
     "_parse_name",
@@ -2326,357 +2323,6 @@ def _frequency(coef: np.ndarray, **kwargs):
     # convert to radians per second
     omega = 2.0 * np.pi * fd / 360.0
     return np.abs(omega)
-
-
-def _love_numbers(
-    omega: np.ndarray,
-    model: str = "PREM",
-    **kwargs,
-):
-    """
-    Compute the body tide Love/Shida numbers for a given frequency
-    :cite:p:`Wahr:1979vx,Wahr:1981ea,Wahr:1981if,Mathews:1995go`
-
-    Parameters
-    ----------
-    omega: np.ndarray
-        Angular frequency (radians per second)
-    model: str, default 'PREM'
-        Earth model to use for Love numbers
-
-            - '1066A'
-            - '1066A-N' (neutral)
-            - '1066A-S' (stable)
-            - 'PEM-C'
-            - 'C2'
-            - 'PREM'
-    astype: np.dtype, default np.float64
-        Data type for the output Love numbers
-
-    Returns
-    -------
-    h2: float
-        Degree-2 Love number of vertical displacement
-    k2: float
-        Degree-2 Love number of gravitational potential
-    l2: float
-        Degree-2 Love (Shida) number of horizontal displacement
-    """
-    # set default keyword arguments
-    kwargs.setdefault("astype", np.float64)
-    # free core nutation frequencies (cycles per sidereal day) and
-    # Love number parameters from Wahr (1981) table 6
-    # and Mathews et al. (1995) table 3
-    if model == "1066A":
-        lambda_fcn = 1.0021714
-        h0, h1 = np.array([6.03e-1, -2.46e-3])
-        k0, k1 = np.array([2.98e-1, -1.23e-3])
-        l0, l1 = np.array([8.42e-2, 7.81e-5])
-    elif model == "1066A-N":
-        lambda_fcn = 1.0021716
-        h0, h1 = np.array([6.03e-1, -2.46e-3])
-        k0, k1 = np.array([2.98e-1, -1.24e-3])
-        l0, l1 = np.array([8.42e-2, 7.82e-5])
-    elif model == "1066A-S":
-        lambda_fcn = 1.0021708
-        h0, h1 = np.array([6.03e-1, -2.46e-3])
-        k0, k1 = np.array([2.98e-1, -1.24e-3])
-        l0, l1 = np.array([8.42e-2, 7.83e-5])
-    elif model == "PEM-C":
-        lambda_fcn = 1.0021771
-        h0, h1 = np.array([6.02e-1, -2.46e-3])
-        k0, k1 = np.array([2.98e-1, -1.24e-3])
-        l0, l1 = np.array([8.39e-2, 7.69e-5])
-    elif model == "C2":
-        lambda_fcn = 1.0021844
-        h0, h1 = np.array([6.02e-1, -2.45e-3])
-        k0, k1 = np.array([2.98e-1, -1.23e-3])
-        l0, l1 = np.array([8.46e-2, 7.58e-5])
-    elif model == "PREM":
-        lambda_fcn = 1.0023214
-        h0, h1 = np.array([5.994e-1, -2.532e-3])
-        k0, k1 = np.array([2.962e-1, -1.271e-3])
-        l0, l1 = np.array([8.378e-2, 7.932e-5])
-    else:
-        raise ValueError(f"Unknown Earth model: {model}")
-    # Love numbers for different frequency bands
-    if omega > 1e-4:
-        # tides in the semi-diurnal band
-        h2 = 0.609
-        k2 = 0.302
-        l2 = 0.0852
-    elif omega < 2e-5:
-        # tides in the long period band
-        h2 = 0.606
-        k2 = 0.299
-        l2 = 0.0840
-    else:
-        # calculate love numbers following J. Wahr (1979)
-        # use resonance formula for tides in the diurnal band
-        # frequency of the o1 tides (radians/second)
-        (omega_o1,) = frequency("o1", **kwargs)
-        # convert frequency from cycles per sidereal day
-        # frequency of free core nutation (radians/second)
-        omega_fcn = lambda_fcn * 7292115e-11
-        # Love numbers for frequency using equation 4.18 of Wahr (1981)
-        # (simplification to use only the free core nutation term)
-        ratio = (omega - omega_o1) / (omega_fcn - omega)
-        h2 = h0 + h1 * ratio
-        k2 = k0 + k1 * ratio
-        l2 = l0 + l1 * ratio
-    # return the Love numbers for frequency
-    return np.array([h2, k2, l2], dtype=kwargs["astype"])
-
-
-def _complex_love_numbers(omega: np.ndarray, **kwargs):
-    """
-    Compute the complex body tide Love/Shida numbers with in-phase
-    and out-of-phase components for a given frequency
-    :cite:p:`Mathews:1997js,Mathews:2002cr,Petit:2010tp`
-
-    Parameters
-    ----------
-    omega: np.ndarray
-        Angular frequency (radians per second)
-    kwargs: dict
-        Keyword arguments for :py:func:`pyTMD.constituents._love_numbers`
-
-    Returns
-    -------
-    h2: complex
-        Degree-2 Love number of vertical displacement
-    k2: complex
-        Degree-2 Love number of gravitational potential
-    l2: complex
-        Degree-2 Love (Shida) number of horizontal displacement
-    """
-    # number of sidereal days per solar day
-    sidereal_ratio = 1.002737909
-    # number of seconds in a sidereal day (approximately 86164.1)
-    sidereal_day = 86400.0 / sidereal_ratio
-    # frequency in cycles per sidereal day
-    f = omega * sidereal_day / (2.0 * np.pi)
-    # Love numbers for different frequency bands
-    if omega == 0.0:
-        # use real-valued body tide Love numbers for the permanent tide
-        # to prevent singularities in frequency-dependent model
-        h2, k2, l2 = _love_numbers(omega, **kwargs)
-    elif omega > 1e-4:
-        # in-phase and out-of-phase components for the semi-diurnal band
-        # table 7.3a (IERS conventions 2010)
-        h2 = 0.6078 - 0.0022j
-        # table 6.5c (IERS conventions 2010)
-        k2 = 0.30102 - 0.0013j
-        # table 7.3a (IERS conventions 2010)
-        l2 = 0.0847 - 0.0007j
-    elif omega < 2e-5:
-        # compute in-phase and out-of-phase components for the long period band
-        # variation largely due to mantle anelasticity
-        alpha = 0.15
-        # frequency equivalent to 200s
-        fm = sidereal_day / 200.0
-        factor = np.tan(alpha * np.pi / 2.0) ** (-1)
-        anelasticity_model = (
-            factor * (1.0 - (fm / f) ** alpha) + 1j * (fm / f) ** alpha
-        )
-        # model for the variation of Love numbers across the zonal tide band
-        # equation 7.4a (IERS conventions 2010)
-        h2 = 0.5998 - 9.96e-4 * anelasticity_model
-        # equation 6.12 (IERS conventions 2010)
-        k2 = 0.29525 - 5.796e-4 * anelasticity_model
-        # equation 7.4b (IERS conventions 2010)
-        l2 = 0.0831 - 3.01e-4 * anelasticity_model
-    else:
-        # in-phase and out-of-phase components for the diurnal band
-        # following IERS conventions and Mathews et al. (2002)
-        # values from equation 6.10 of IERS conventions 2010
-        # and from Mathews et al. (2002)
-        sigma = np.zeros((4), dtype=np.complex128)
-        # factor for calculating L0
-        sigma[0] = f - 1.0
-        # Chandler wobble
-        sigma[1] = -0.0026010 - 0.0001361j
-        # retrograde free core nutation
-        sigma[2] = 1.0023181 + 0.000025j
-        # prograde free core nutation
-        sigma[3] = 0.999026 + 0.000780j
-        # frequency dependence of Love number h2 (vertical)
-        # table 7.1 (IERS conventions 2010)
-        H2 = np.zeros((4), dtype=np.complex128)
-        H2[0] = 0.60671 - 0.242e-2j
-        H2[1] = -0.15777e-2 - 0.7630e-4j
-        H2[2] = 0.18053e-3 - 0.6292e-5j
-        H2[3] = -0.18616e-5 + 0.1379e-6j
-        # frequency dependence of Love number k2 (potential)
-        # table 6.4 (IERS conventions 2010)
-        K2 = np.zeros((4), dtype=np.complex128)
-        K2[0] = 0.29954 - 0.1412e-2j
-        K2[1] = -0.77896e-3 - 0.3711e-4j
-        K2[2] = 0.90963e-4 - 0.2963e-5j
-        K2[3] = -0.11416e-5 + 0.5325e-7j
-        # frequency dependence of Love number l2 (horizontal)
-        # table 7.1 (IERS conventions 2010)
-        L2 = np.zeros((4), dtype=np.complex128)
-        L2[0] = 0.84963e-1 - 0.7395e-3j
-        L2[1] = -0.22107e-3 - 0.9646e-5j
-        L2[2] = -0.54710e-5 - 0.2990e-6j
-        L2[3] = -0.29904e-7 - 0.7717e-8j
-        # estimate the complex Love numbers for diurnal tides
-        # equation 6.9 (IERS conventions 2010)
-        h2 = np.sum(H2 / (f - sigma))
-        k2 = np.sum(K2 / (f - sigma))
-        l2 = np.sum(L2 / (f - sigma))
-
-    # return the Love numbers as a complex number
-    # to include the in-phase and out-of-phase components
-    return np.array([h2, k2, l2], dtype=np.complex128)
-
-
-def _degree_love_numbers(
-    l: int,
-    model: str = "Longman",
-    **kwargs,
-):
-    """
-    Extracts body tide Love/Shida numbers for a given degree
-    :cite:p:`Melchior:1983wd`
-
-    Parameters
-    ----------
-    l: int
-        Degree of the spherical harmonics
-    model: str, default "Longman"
-        Earth model
-
-        - ``'Longman'``: :cite:t:`Longman:1959hw`
-        - ``'Kaula'``: :cite:t:`Kaula:1966un`
-        - ``'Takeuchi'``: :cite:t:`Takeuchi:1950hi`
-
-    Returns
-    -------
-    hl: float
-        Body tide Love number of vertical displacement
-    kl: float
-        Body tide Love number of gravitational potential
-    ll: float
-        Body tide Love (Shida) number of horizontal displacement
-    """
-    # get the table of body tide love numbers
-    table = _melchior_table_52(model)
-    # provide zero values for degrees not provided in the table
-    lmin, lmax = np.array([table[0, 0], table[-1, 0]], dtype="i")
-    if (l < lmin) | (l > lmax):
-        return (0.0, 0.0, 0.0)
-    # column 1: Spherical harmonic degree
-    n = table[l - lmin, 0]
-    # verify that the rows match
-    if n != l:
-        raise ValueError(f"Mismatched row for degree {l}")
-    # column 2: Love number of vertical displacement
-    hl = table[l - lmin, 1]
-    # column 3: Love number of gravitational potential
-    kl = table[l - lmin, 2]
-    # column 4: Shida number of horizontal displacement
-    ll = table[l - lmin, 3]
-    # return the Love numbers for degree l
-    return (hl, kl, ll)
-
-
-def _melchior_table_52(model: str):
-    """
-    Body tide Love numbers for an Earth model provided in
-    Table 5.2 of :cite:t:`Melchior:1983wd`
-
-    Parameters
-    ----------
-    model: str
-        Earth model
-
-        - ``'Longman'``: :cite:t:`Longman:1959hw`
-        - ``'Kaula'``: :cite:t:`Kaula:1966un`
-        - ``'Takeuchi'``: :cite:t:`Takeuchi:1950hi`
-    """
-    # table 5.2 from Melchior (1983)
-    # column 1: Spherical harmonic degree
-    # column 2: Love number of vertical displacement
-    # column 3: Love number of gravitational potential
-    # column 4: Shida number of horizontal displacement
-    if model == "Longman":
-        # values from Longman (1959)
-        table_52 = np.array(
-            [
-                [2, 0.612, 0.302, 0.083],
-                [3, 0.290, 0.093, 0.014],
-                [4, 0.175, 0.042, 0.010],
-                [5, 0.129, 0.025, 0.008],
-                [6, 0.107, 0.017, 0.007],
-                [7, 0.095, 0.013, 0.005],
-                [8, 0.087, 0.010, 0.004],
-                [9, 0.081, 0.008, 0.004],
-                [10, 0.076, 0.007, 0.003],
-                [11, 0.072, 0.006, 0.002],
-                [12, 0.069, 0.005, 0.002],
-                [13, 0.066, 0.005, 0.002],
-                [14, 0.064, 0.004, 0.001],
-                [15, 0.062, 0.004, 0.001],
-                [16, 0.060, 0.003, 0.001],
-                [17, 0.058, 0.003, 0.001],
-                [18, 0.056, 0.003, 0.001],
-                [19, 0.055, 0.003, 0.001],
-                [20, 0.053, 0.002, 0.001],
-                [21, 0.052, 0.002, 0.001],
-                [22, 0.051, 0.002, 0.000],
-                [23, 0.050, 0.002, 0.000],
-                [24, 0.048, 0.002, 0.000],
-                [25, 0.047, 0.002, 0.000],
-            ]
-        )
-    elif model == "Kaula":
-        # values from Kaula (1966)
-        table_52 = np.array(
-            [
-                [2, 0.624, 0.317, 0.085],
-                [3, 0.293, 0.095, 0.014],
-                [4, 0.177, 0.042, 0.010],
-                [5, 0.130, 0.025, 0.008],
-                [6, 0.107, 0.017, 0.007],
-                [7, 0.095, 0.013, 0.005],
-                [8, 0.087, 0.010, 0.004],
-                [9, 0.081, 0.008, 0.004],
-                [10, 0.076, 0.007, 0.003],
-                [11, 0.072, 0.006, 0.002],
-                [12, 0.069, 0.005, 0.002],
-                [13, 0.066, 0.005, 0.002],
-                [14, 0.064, 0.004, 0.001],
-                [15, 0.062, 0.004, 0.001],
-                [16, 0.060, 0.003, 0.001],
-            ]
-        )
-    elif model == "Takeuchi":
-        # values from Takeuchi (1950)
-        table_52 = np.array(
-            [
-                [2, 0.592, 0.280, 0.076],
-                [3, 0.274, 0.083, 0.010],
-                [4, 0.161, 0.035, 0.007],
-                [5, 0.116, 0.020, 0.006],
-                [6, 0.094, 0.013, 0.005],
-                [7, 0.081, 0.009, 0.004],
-                [8, 0.073, 0.007, 0.003],
-                [9, 0.068, 0.006, 0.0025],
-                [10, 0.063, 0.005, 0.002],
-                [11, 0.059, 0.004, 0.002],
-                [12, 0.055, 0.003, 0.002],
-                [13, 0.053, 0.003, 0.0015],
-                [14, 0.051, 0.003, 0.001],
-                [15, 0.0495, 0.0025, 0.001],
-                [16, 0.048, 0.002, 0.001],
-            ]
-        )
-    else:
-        raise ValueError(f"Unknown Earth model: {model}")
-    # return the table of love numbers
-    return table_52
 
 
 # Doodson (1921) table with values missing from Cartwright tables
